DR-465 gcs profile aware (#220)
* redoing after a bad attempt at refactoring

* making progress

* intermediate

* intermediate

* gcs pdao refactored

* gets tests working

* remove FileData*

* getting tests working

* missed Storage @Autowire

* fixes integration tests

+++ src/main/java/bio/terra/pdao/gcs/GcsPdao.java
@@ -1,14 +1,19 @@
 package bio.terra.pdao.gcs;
 
+import bio.terra.filesystem.FireStoreFileDao;
 import bio.terra.metadata.FSFile;
 import bio.terra.metadata.FSFileInfo;
 import bio.terra.metadata.Dataset;
+import bio.terra.metadata.FSObjectBase;
+import bio.terra.metadata.FSObjectType;
 import bio.terra.model.FileLoadModel;
 import bio.terra.pdao.exception.PdaoException;
 import bio.terra.pdao.exception.PdaoFileCopyException;
 import bio.terra.pdao.exception.PdaoInvalidUriException;
 import bio.terra.pdao.exception.PdaoSourceFileNotFoundException;
-import com.google.api.gax.paging.Page;
+import bio.terra.resourcemanagement.metadata.google.GoogleBucketResource;
+import bio.terra.resourcemanagement.metadata.google.GoogleProjectResource;
+import bio.terra.service.dataproject.DataLocationService;
 import com.google.cloud.storage.Acl;
 import com.google.cloud.storage.Blob;
 import com.google.cloud.storage.BlobId;
@@ -16,6 +21,8 @@ import com.google.cloud.storage.CopyWriter;
 import com.google.cloud.storage.Storage;
 import com.google.cloud.storage.StorageException;
 import org.apache.commons.lang3.StringUtils;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
@@ -24,24 +31,39 @@ import java.net.URI;
 import java.net.URISyntaxException;
 import java.time.Instant;
 import java.util.List;
+import java.util.Optional;
+import java.util.UUID;
 
 @Component
 @Profile("google")
 public class GcsPdao {
-    private GcsConfiguration gcsConfiguration;
-    private Storage storage;
+    private static final Logger logger = LoggerFactory.getLogger(GcsPdao.class);
+    private static final int DATASET_DELETE_BATCH_SIZE = 1000;
+
+    private final GcsProjectFactory gcsProjectFactory;
+    private final DataLocationService dataLocationService;
+    private final FireStoreFileDao fileDao;
 
     @Autowired
-    public GcsPdao(GcsConfiguration gcsConfiguration, Storage storage) {
-        this.gcsConfiguration = gcsConfiguration;
-        this.storage = storage;
+    public GcsPdao(
+        GcsProjectFactory gcsProjectFactory,
+        DataLocationService dataLocationService,
+        FireStoreFileDao fileDao) {
+        this.gcsProjectFactory = gcsProjectFactory;
+        this.dataLocationService = dataLocationService;
+        this.fileDao = fileDao;
+    }
+
+    private Storage storageForBucket(GoogleBucketResource bucketResource) {
+        GoogleProjectResource projectResource = bucketResource.getProjectResource();
+        GcsProject gcsProject = gcsProjectFactory.get(projectResource.getGoogleProjectId());
+        return gcsProject.getStorage();
     }
 
     // We return the incoming FSObject with the blob information filled in
     public FSFileInfo copyFile(Dataset dataset,
                                FileLoadModel fileLoadModel,
                                FSFile fsObject) {
-
         String sourceBucket;
         String sourcePath;
 
@@ -54,7 +76,6 @@ public class GcsPdao {
             if (sourceUri.getPort() != -1) {
                 throw new PdaoInvalidUriException("Source path must not have a port specified: '" +
                     fileLoadModel.getSourcePath() + "'");
-
             }
             sourceBucket = sourceUri.getAuthority();
             sourcePath = StringUtils.removeStart(sourceUri.getPath(), "/");
@@ -67,6 +88,12 @@ public class GcsPdao {
             throw new PdaoInvalidUriException("Invalid gs path: '" +
                 fileLoadModel.getSourcePath() + "'");
         }
+
+        // Figure out where the destination bucket is. Similar to how there is a ProjectIdSelector there will be need
+        // to be a BucketSelector that picks a (Project, Bucket) pair for the file. The Resource Manager will then need
+        // to potentially create the project and/or the bucket inside of that project depending on what already exists.
+        GoogleBucketResource bucketForFile = dataLocationService.getBucketForFile(fsObject);
+        Storage storage = storageForBucket(bucketForFile);
 
         Blob sourceBlob = storage.get(BlobId.of(sourceBucket, sourcePath));
         if (sourceBlob == null) {
@@ -84,7 +111,7 @@ public class GcsPdao {
             // I have been seeing timeouts and I think they are due to particularly large files,
             // so I changed exported the timeouts to application.properties to allow for tuning
             // and I am changing this to copy chunks.
-            CopyWriter writer = sourceBlob.copyTo(BlobId.of(gcsConfiguration.getBucket(), targetPath));
+            CopyWriter writer = sourceBlob.copyTo(BlobId.of(bucketForFile.getName(), targetPath));
             while (!writer.isDone()) {
                 writer.copyChunk();
             }
@@ -105,7 +132,7 @@ public class GcsPdao {
             Instant createTime = Instant.ofEpochMilli(targetBlob.getCreateTime());
 
             URI gspath = new URI("gs",
-                gcsConfiguration.getBucket(),
+                bucketForFile.getName(),
                 "/" + targetPath,
                 null,
                 null);
@@ -117,7 +144,9 @@ public class GcsPdao {
                 .gspath(gspath.toString())
                 .checksumCrc32c(targetBlob.getCrc32cToHexString())
                 .checksumMd5(checksumMd5)
-                .size(targetBlob.getSize());
+                .size(targetBlob.getSize())
+                .region(bucketForFile.getRegion())
+                .bucketResourceId(bucketForFile.getResourceId().toString());
 
             return fsFileInfo;
         } catch (StorageException ex) {
@@ -130,27 +159,29 @@ public class GcsPdao {
         }
     }
 
-    public boolean deleteFile(Dataset dataset, String fileId) {
-        String dataBucket = gcsConfiguration.getBucket();
-        String filePath = dataset.getId().toString() + "/" + fileId;
-
-        Blob blob = storage.get(BlobId.of(dataBucket, filePath));
-        if (blob == null) {
-            return false;
+    public boolean deleteFile(FSFile fsFile) {
+        GoogleBucketResource bucketForFile = dataLocationService.getBucketForFile(fsFile);
+        Storage storage = storageForBucket(bucketForFile);
+        // It's possible that the file didn't get written to a bucket, meaning gspath will be null.
+        Optional<String> gspath = Optional.ofNullable(fsFile.getGspath());
+        if (gspath.isPresent()) {
+            URI uri = URI.create(gspath.get());
+            String bucketPath = StringUtils.removeStart(uri.getPath(), "/");
+            Optional<Blob> blob = Optional.ofNullable(storage.get(BlobId.of(bucketForFile.getName(), bucketPath)));
+            if (blob.isPresent()) {
+                return blob.get().delete();
+            }
         }
-        return blob.delete();
+        return false;
     }
 
     public void deleteFilesFromDataset(Dataset dataset) {
-        String dataBucket = gcsConfiguration.getBucket();
-        String directory = dataset.getId().toString() + "/";
-        Page<Blob> blobs = storage.list(
-            dataBucket,
-            Storage.BlobListOption.currentDirectory(),
-            Storage.BlobListOption.prefix(directory));
-        for (Blob blob : blobs.iterateAll()) {
-            blob.delete();
-        }
+        // these files could be in multiple buckets..need to enumerate them from firestore and then iterate + delete
+        fileDao.forEachFsObjectInDataset(dataset, DATASET_DELETE_BATCH_SIZE, fsObjectBase -> {
+            if (fsObjectBase.getObjectType() != FSObjectType.DIRECTORY) {
+                deleteFile((FSFile) fsObjectBase);
+            }
+        });
     }
 
     private enum AclOp {
@@ -158,29 +189,35 @@ public class GcsPdao {
         ACL_OP_DELETE
     };
 
-    public void setAclOnFiles(String datasetId, List<String> fileIds, String readersPolicyEmail) {
-        fileAclOp(AclOp.ACL_OP_CREATE, datasetId, fileIds, readersPolicyEmail);
+    public void setAclOnFiles(Dataset dataset, List<String> fileIds, String readersPolicyEmail) {
+        fileAclOp(AclOp.ACL_OP_CREATE, dataset, fileIds, readersPolicyEmail);
     }
 
-    public void removeAclOnFiles(String datasetId, List<String> fileIds, String readersPolicyEmail) {
-        fileAclOp(AclOp.ACL_OP_DELETE, datasetId, fileIds, readersPolicyEmail);
+    public void removeAclOnFiles(Dataset dataset, List<String> fileIds, String readersPolicyEmail) {
+        fileAclOp(AclOp.ACL_OP_DELETE, dataset, fileIds, readersPolicyEmail);
     }
 
-    private void fileAclOp(AclOp op, String datasetId, List<String> fileIds, String readersPolicyEmail) {
-        String dataBucket = gcsConfiguration.getBucket();
-        String directory = datasetId + "/";
+    private void fileAclOp(AclOp op, Dataset dataset, List<String> fileIds, String readersPolicyEmail) {
         Acl.Group readerGroup = new Acl.Group(readersPolicyEmail);
         Acl acl = Acl.newBuilder(readerGroup, Acl.Role.READER).build();
+
         for (String fileId : fileIds) {
-            String blobName = directory + fileId;
-            BlobId blobId = BlobId.of(dataBucket, blobName);
-            switch (op) {
-                case ACL_OP_CREATE:
-                    storage.createAcl(blobId, acl);
-                    break;
-                case ACL_OP_DELETE:
-                    storage.deleteAcl(blobId, readerGroup);
-                    break;
+            FSObjectBase fsObjectBase = fileDao.retrieve(dataset, UUID.fromString(fileId));
+            if (fsObjectBase.getObjectType() == FSObjectType.FILE) {
+                FSFile fsFile = (FSFile)fsObjectBase;
+                GoogleBucketResource bucketForFile = dataLocationService.getBucketForFile(fsFile);
+                Storage storage = storageForBucket(bucketForFile);
+                URI gsUri = URI.create(fsFile.getGspath());
+                String bucketPath = StringUtils.removeStart(gsUri.getPath(), "/");
+                BlobId blobId = BlobId.of(bucketForFile.getName(), bucketPath);
+                switch (op) {
+                    case ACL_OP_CREATE:
+                        storage.createAcl(blobId, acl);
+                        break;
+                    case ACL_OP_DELETE:
+                        storage.deleteAcl(blobId, readerGroup);
+                        break;
+                }
             }
         }
     }
